--- app/drag-n-drop/index.js
@@ -1,2 +1,0 @@
-
-module.exports = {};
